Create Photo parser
The script performs the following tasks:

Extracts EXIF data from an image file using PIL.
Creates a Google Earth KML file based on the GPS coordinates found in the EXIF data.
Prints the URL to view the KML data in a browser.
It takes two command line arguments: the path to the input image file and the path to the output KML file. The script uses the argparse library to parse these arguments and make them available to the main function.

The extract_exif function takes an image path as an argument and returns a dictionary of EXIF data for that image. The create_kml function takes the EXIF data and a KML file path as arguments, and creates a KML file based on the GPS coordinates found in the EXIF data. If GPS coordinates are found, the URL to view the KML data in a browser is returned.

Finally, the main function uses the extract_exif and create_kml functions to extract EXIF data from the input image and create a KML file if GPS coordinates are found. The URL to view the KML data in a browser is printed if the KML file is created successfully, otherwise a message is printed indicating that no GPS coordinates were found in the image.